ft_substr 1.0

+++ libft/Part2/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: aymaatou <aymaatou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 16:25:15 by aymaatou          #+#    #+#             */
-/*   Updated: 2019/10/20 16:21:34 by aymaatou         ###   ########.fr       */
+/*   Updated: 2019/10/20 17:09:06 by aymaatou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 static char	*f_trim(char *l_ptr, const char *set)
 {
 	size_t	size_set;
-	int		i;
+	size_t		i;
 
 	size_set = ft_strlen(set);
 	i = 0;
@@ -36,7 +36,7 @@ static char	*f_trim(char *l_ptr, const char *set)
 static char	*b_trim(char *f_ptr, const char *set)
 {
 	size_t	size_set;
-	int		i;
+	size_t		i;
 
 	size_set = ft_strlen(set);
 	i = 0;
@@ -58,8 +58,8 @@ char		*ft_strtrim(char const *s1, char const *set)
 	char	*l_ptr;
 	char	*f_ptr;
 	char	*re_str;
-	int		i;
-	int		re_length;
+	size_t		i;
+	size_t		re_length;
 
 	l_ptr = (char *)s1 + ft_strlen(s1) - 1;
 	f_ptr = (char *)s1;
